Cleanup of claimable contract
Removed the preclaim. The preclaim did not add any protection in cases where the tokens got sent to an invalid address. And in the case of a valid address, frontrunning a claim is very risky as the claimer might be "fishing" for the frontrunning and then take the collateral.

Also using uint256 everywhere.

Binaries that are claimable are now about 4 kB smaller, which is good.